HematoVision/requirements.txt
tensorflow==2.16.2
Flask==3.0.3
numpy==1.26.4
scikit-learn==1.4.2
Pillow==10.3.0
